Warning fix

+++ src/tools/rocketbench/main.cc
@@ -622,7 +622,7 @@ static int SaveFile(std::string filename,
   if (fclose(fh)) {
     return errno;
   }
-  if (FLAGS_num_messages != num_messages) {
+  if (static_cast<uint64_t>(FLAGS_num_messages) != num_messages) {
     LOG_ERROR(info_log, "Number of messages produced %zu does not match "
               " number of messages saved in file %zu\n",
               FLAGS_num_messages,
